Add comments, rearrange a bit of code
