+++ CourseGrab/Extensions/UIImage+Shared.swift
@@ -15,4 +15,22 @@ extension UIImage {
     static let searchIcon = UIImage(named: "icon_search")
     static let settingsIcon = UIImage(named: "icon_settings")
 
+    func with(insets: UIEdgeInsets) -> UIImage? {
+        UIGraphicsBeginImageContextWithOptions(
+            CGSize(
+                width: size.width + insets.left + insets.right,
+                height: size.height + insets.top + insets.bottom
+            ),
+            false,
+            self.scale
+        )
+
+        let origin = CGPoint(x: insets.left, y: insets.top)
+        self.draw(at: origin)
+        let imageWithInsets = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+
+        return imageWithInsets
+    }
+
 }
