+++ CourseGrab/Home/HomeViewController.swift
@@ -7,7 +7,7 @@
 //
 
 import UIKit
-import SnapKit
+import Tactile
 
 class HomeViewController: UITableViewController {
 
@@ -47,8 +47,8 @@ class HomeViewController: UITableViewController {
             )
         ]
 
-        view.backgroundColor = .white
-
+        // Setup navigation bar
+        
         let titleLabel = UILabel()
         titleLabel.text = "CourseGrab"
         titleLabel.textColor = .white
@@ -61,9 +61,12 @@ class HomeViewController: UITableViewController {
 
         let searchButton = UIButton(type: .system)
         searchButton.setImage(.searchIcon, for: .normal)
-        searchButton.addTarget(self, action: #selector(showSearch), for: .touchUpInside)
+        searchButton.on(.touchUpInside, showSearch)
         navigationItem.rightBarButtonItem = UIBarButtonItem(customView: searchButton)
 
+        // Setup tableView
+        
+        tableView.backgroundColor = .white
         tableView.separatorStyle = .none
         tableView.rowHeight = UITableView.automaticDimension
         tableView.register(HomeTableViewCell.self, forCellReuseIdentifier: homeCellReuseId)
@@ -91,7 +94,8 @@ extension HomeViewController {
 
 extension HomeViewController {
 
-    @objc func showSearch() {
+    private func showSearch(_ button: UIButton) {
+        // Grab navigation bar views and frames
         guard let navigationBar = navigationController?.navigationBar,
             let leftButton = navigationItem.leftBarButtonItem?.customView as? UIButton,
             let leftButtonFrame = leftButton.superview?.convert(leftButton.frame, to: navigationBar),
@@ -101,10 +105,13 @@ extension HomeViewController {
             let titleLabelFrame = titleLabel.superview?.convert(titleLabel.frame, to: navigationBar)
             else { return }
 
+        // Remove views from navigationBar
         navigationItem.leftBarButtonItem = nil
         navigationItem.rightBarButtonItem = nil
         navigationItem.titleView = nil
 
+        // Add views back to navigationBar but so that they aren't
+        // controlled by autolayout and we can animate them.
         leftButton.frame = leftButtonFrame
         navigationBar.addSubview(leftButton)
         rightButton.frame = rightButtonFrame
@@ -112,10 +119,16 @@ extension HomeViewController {
         titleLabel.frame = titleLabelFrame
         navigationBar.addSubview(titleLabel)
 
+        // Update the search button to look like a back button. We
+        // have to update the frame because the images are not the
+        // same size.
         rightButton.setImage(.backIcon, for: .normal)
         rightButton.sizeToFit()
         rightButton.frame.origin.y = rightButtonFrame.origin.y + (rightButtonFrame.height - rightButton.frame.height) / 2
 
+        // Create the textField and transform it so it is a little
+        // off from the final destination. We will animate the
+        // transform away for a swiping effect.
         let textField = UITextField()
         textField.attributedPlaceholder = NSAttributedString(
             string: "Search for a course",
@@ -134,6 +147,9 @@ extension HomeViewController {
         textField.center.y = leftButton.center.y
         textField.transform = CGAffineTransform(translationX: 20, y: 0)
 
+        // Perform the animations, where we hide the views we
+        // don't want anymore and transform them to create the
+        // swiping effect.
         UIView.animate(withDuration: 0.3, animations: {
             leftButton.alpha = 0
             leftButton.transform = CGAffineTransform(translationX: -20, y: 0)
@@ -146,23 +162,24 @@ extension HomeViewController {
             textField.alpha = 1
             textField.transform = .identity
         }) { _ in
+            // Remove the views from the navigationBar. We don't
+            // want them staying there!
             leftButton.removeFromSuperview()
             rightButton.removeFromSuperview()
             titleLabel.removeFromSuperview()
             textField.removeFromSuperview()
 
+            // Remove the transform from the right button so it
+            // doesn't affect us when we don't expect it.
             rightButton.transform = .identity
-
-            self.navigationItem.leftBarButtonItems = [
-                UIBarButtonItem(customView: rightButton),
-                UIBarButtonItem(customView: textField)
-            ]
         }
 
+        // Create and present the SearchViewController by fading it in.
         let searchViewController = SearchViewController()
         UIView.transition(from: view, to: searchViewController.view, duration: 0.3, options: .transitionCrossDissolve) { _ in
             self.navigationController?.pushViewController(searchViewController, animated: false)
 
+            // Reset the navigationBar for when the user swipes back.
             leftButton.alpha = 1
             rightButton.setImage(.searchIcon, for: .normal)
             rightButton.sizeToFit()
